pyproject.toml
[project]
name = "NetWorkToolsBot"
version = "0.1.0"
description = "NetWorkToolsBot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot2[httpx]>=2.4.4",
    "nonebot2[websockets]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.60.4",
    "dnspython>=2.8.0",
]

[tool.nonebot]
plugin_dirs = ["./src/plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []


[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-alconna = ["nonebot_plugin_alconna"]
